Change SetJSON to correctly handle pointer args
This fixes an issue with not grabbed `goriakindex` tags when we pass the pointer to SetJSON method.

Before this commit the behavior was:
- call SetJSON with the pointer to a struct which has `goriakindex` tag.
- under the hood, the method determinates kind of passed value as `ptr` and doesn't try to find the field with tag.
- data are silently written into a bucket, without a secondary key.

In this commit added an extra check on pointer and data passed by pointer handled in the same way as data passed by value.
Test case included into the commit.